+++ examples/shared_samples/logical_interconnect.rb
@@ -1,4 +1,4 @@
-# (C) Copyright 2017 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2020 Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # You may not use this file except in compliance with the License.
@@ -14,16 +14,9 @@ require_relative '../_client' # Gives access to @client
 # Example: Explores functionalities of Logical Interconnects
 #
 # Supported APIs:
-# - 200, 300, 500, 600
-
-# Resources that can be created according to parameters:
-# api_version = 200 & variant = any to OneviewSDK::API200::LogicalInterconnect
-# api_version = 300 & variant = C7000 to OneviewSDK::API300::C7000::LogicalInterconnect
-# api_version = 300 & variant = Synergy to OneviewSDK::API300::Synergy::LogicalInterconnect
-# api_version = 500 & variant = C7000 to OneviewSDK::API500::C7000::LogicalInterconnect
-# api_version = 500 & variant = Synergy to OneviewSDK::API500::Synergy::LogicalInterconnect
-# api_version = 600 & variant = C7000 to OneviewSDK::API600::C7000::LogicalInterconnect
-# api_version = 600 & variant = Synergy to OneviewSDK::API600::Synergy::LogicalInterconnect
+# - 200, 300, 500, 600, 800, 1000, 1200
+# Supported variants:
+# - C7000, Synergy
 
 # Resource Class used in this sample
 logical_interconnect_class = OneviewSDK.resource_named('LogicalInterconnect', @client.api_version)